refactor(initialroutenamestate): simplify API
To be completely honest, there's no need for the React app to know when to change the initial route state because that's already done at the bridging layer. At this layer, we preserve the integrity of the initial route name state, such that only navigation-related pieces (MainActivity and InitialRouteNameStateContext) need to ever update it.
