+++ src/App.tsx
@@ -12,14 +12,22 @@ import { NavigationContainer } from '@react-navigation/native';
 import React from 'react';
 
 import Flow from './flow/ui/Flow';
+import { getInitialRouteNameStateContext } from './navigation/initialRouteNameState/InitialRouteNameStateContext';
+import useInitialRouteNameStateContextProviderValue from './navigation/initialRouteNameState/useInitialRouteNameStateContextProviderValue';
 
 export const APP_NAME = 'AndroidStartupPerfApp';
 
+const InitialRouteNameStateContext = getInitialRouteNameStateContext();
+
 const App = () => {
+  const initialRouteNameStateContextValue = useInitialRouteNameStateContextProviderValue();
+
   return (
-    <NavigationContainer>
-      <Flow />
-    </NavigationContainer>
+    <InitialRouteNameStateContext.Provider value={initialRouteNameStateContextValue}>
+      <NavigationContainer>
+        <Flow />
+      </NavigationContainer>
+    </InitialRouteNameStateContext.Provider>
   );
 };
